requirements.txt
opencv-python>=4.8
Pillow>=10.0
numpy>=1.24
scikit-learn>=1.3
face-recognition>=1.3
